+++ ui/src/store/camview.module.js
@@ -1,5 +1,5 @@
-const layout = JSON.parse(localStorage.getItem('camview-layout'));
-const initialState = { layout: layout || [] };
+const layout = JSON.parse(localStorage.getItem('camview-layouts'));
+const initialState = { layout: layout || {} };
 
 export const camview = {
   namespaced: true,
@@ -12,8 +12,10 @@ export const camview = {
   },
   mutations: {
     updateElements: (state, payload) => {
-      localStorage.setItem('camview-layout', JSON.stringify(payload));
-      state.layout = payload;
+      const storedLayout = JSON.parse(localStorage.getItem('camview-layouts')) || {};
+      storedLayout[Object.keys(payload)[0]] = Object.values(payload)[0];
+      localStorage.setItem('camview-layouts', JSON.stringify(storedLayout));
+      state.layout = storedLayout;
     },
   },
 };