+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Interface:
   - Added `auto darkmode` switch to settings to automatically switch darkmode based on system settings (if supported)
 
+## Bugfixes
+- Minor UI bugfixes
+
 # v4.2.3 - 2021-05-06
 
 ## Bugfixes